+++ Test/MoveDepthTest.cs
@@ -11,19 +11,19 @@ namespace Test
     public class MoveDepthTest
     {
 
-        private const int D0 = 1;
-        private const int D1 = 20;
-        private const int D2 = 400;
-        private const int D3 = 8902;
-        private const int D4 = 197281;
-        private const int D5 = 4865609;
-        private const int D6 = 119060324;
+        private const ulong D0 = 1;
+        private const ulong D1 = 20;
+        private const ulong D2 = 400;
+        private const ulong D3 = 8902;
+        private const ulong D4 = 197281;
+        private const ulong D5 = 4865609;
+        private const ulong D6 = 119060324;
         
         private readonly DataBoard Board = DataBoard.Default();
 
         private int SelectedDepth;
 
-        private static void LogNodeCount((int, int) piece, (int, int) move, int nodeC)
+        private static void LogNodeCount((int, int) piece, (int, int) move, ulong nodeC)
         {
             string fullMove = Util.TupleToChessString(piece) + Util.TupleToChessString(move);
             Console.WriteLine(fullMove.ToLower() + ": " + nodeC);
@@ -46,59 +46,59 @@ namespace Test
             MoveGeneration(Board, 1, verbose: false);
         }
         
-        public (int, int) Depth0()
+        public (ulong, ulong) Depth0()
         {
             SelectedDepth = 0;
             return (D0, MoveGeneration(Board, 0));
         }
         
-        public (int, int) Depth1()
+        public (ulong, ulong) Depth1()
         {
             SelectedDepth = 1;
             return (D1, MoveGeneration(Board, 1));
         }
         
-        public (int, int) Depth2()
+        public (ulong, ulong) Depth2()
         {
             SelectedDepth = 2;
             return (D2, MoveGeneration(Board, 2));
         }
 
-        public (int, int) Depth3()
+        public (ulong, ulong) Depth3()
         {
             SelectedDepth = 3;
             return (D3, MoveGeneration(Board, 3));
         }
 
-        public (int, int) Depth4()
+        public (ulong, ulong) Depth4()
         {
             SelectedDepth = 4;
             return (D4, MoveGeneration(Board, 4));
         }
         
-        public (int, int) Depth5()
+        public (ulong, ulong) Depth5()
         {
             SelectedDepth = 5;
             return (D5, MoveGeneration(Board, 5));
         }
         
-        public (int, int) Depth6()
+        public (ulong, ulong) Depth6()
         {
             SelectedDepth = 6;
             return (D6, MoveGeneration(Board, 6));
         }
         
-        private int MoveGeneration(DataBoard board, int depth, PieceColor color = PieceColor.White, bool verbose = true)
+        private ulong MoveGeneration(DataBoard board, int depth, PieceColor color = PieceColor.White, bool verbose = true)
         {
             if (depth == 0) return 1;
-            int count = 0;
+            ulong count = 0;
             
             if (depth == SelectedDepth) color = board.IsWhiteTurn() ? PieceColor.White : PieceColor.Black;
 
             if (depth == 1) {
                 foreach ((int, int) piece in board.All(color)) {
                     LegalMoveSet moveSet = new(board, piece);
-                    count += moveSet.Count();
+                    count += (ulong) moveSet.Count();
                     if (depth != SelectedDepth) continue;
                     
                     if (verbose) LogNodeCount(piece, moveSet.Count());
@@ -114,7 +114,7 @@ namespace Test
                 foreach ((int, int) move in moveSet.Get()) {
                     DataBoard nextBoard = board.Clone();
                     nextBoard.Move(piece, move);
-                    int generatedCount = MoveGeneration(nextBoard, depth - 1, Util.OppositeColor(color));
+                    ulong generatedCount = MoveGeneration(nextBoard, depth - 1, Util.OppositeColor(color));
                     Interlocked.Add(
                         ref count,
                         generatedCount
